backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
firebase-admin==6.5.0
pydantic==2.7.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
tensorflow==2.16.1
joblib==1.4.2
aiohttp==3.9.5
python-multipart==0.0.9
